feat: complete full node implementation with consensus and native contracts
- Implement consensus state persistence and recovery
- Add message caching for replay attack prevention
- Complete block assembly with multi-sig witness construction
- Add StdLib native contract methods (memorySearch, stringSplit, strLen)
- Implement state trie manager for MPT state root calculation
- Add block executor for transaction processing
- Complete validator service with wallet loading
- Add integration tests for all layers
- Clean up outdated documentation

+++ neo-mempool/src/policy.rs
@@ -24,9 +24,9 @@ pub struct FeePolicy {
 impl Default for FeePolicy {
     fn default() -> Self {
         Self {
-            min_fee_per_byte: 1000,         // 0.00001 GAS per byte
-            max_fee_per_byte: 10_000_000,   // 0.1 GAS per byte
-            low_priority_threshold: 100_000, // 0.001 GAS
+            min_fee_per_byte: 1000,              // 0.00001 GAS per byte
+            max_fee_per_byte: 10_000_000,        // 0.1 GAS per byte
+            low_priority_threshold: 100_000,     // 0.001 GAS
             high_priority_threshold: 10_000_000, // 0.1 GAS
             congestion_factor: 1.0,
         }
@@ -51,7 +51,7 @@ impl FeePolicy {
             1000
         } else {
             let range = self.max_fee_per_byte - self.min_fee_per_byte;
-            ((fee_per_byte - self.min_fee_per_byte) * 1000 / range) as i64
+            (fee_per_byte - self.min_fee_per_byte) * 1000 / range
         };
 
         // Apply congestion factor
